requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile .\pyproject.toml --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
asttokens==3.0.0
    # via stack-data
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via
    #   pdfminer-six
    #   requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   ipython
    #   pytest
    #   tqdm
    #   uvicorn
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
cryptography==45.0.5
    # via pdfminer-six
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.2.0
    # via fastapi
executing==2.2.0
    # via stack-data
fastapi==0.115.14
    # via uv-example (./pyproject.toml)
fastapi-cli==0.0.7
    # via fastapi
google-auth==2.40.3
    # via google-genai
google-genai==1.24.0
    # via uv-example (./pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   google-genai
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
ipykernel==6.29.5
    # via uv-example (./pyproject.toml)
ipython==9.4.0
    # via
    #   ipykernel
    #   ipywidgets
ipython-pygments-lexers==1.1.1
    # via ipython
ipywidgets==8.1.7
    # via uv-example (./pyproject.toml)
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via fastapi
jiter==0.10.0
    # via openai
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
jupyterlab-widgets==3.0.15
    # via ipywidgets
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
nest-asyncio==1.6.0
    # via ipykernel
openai==1.93.0
    # via uv-example (./pyproject.toml)
packaging==25.0
    # via
    #   ipykernel
    #   pytest
parso==0.8.4
    # via jedi
pdfminer-six==20250506
    # via pdfplumber
pdfplumber==0.11.7
    # via uv-example (./pyproject.toml)
pillow==11.3.0
    # via pdfplumber
platformdirs==4.3.8
    # via jupyter-core
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.51
    # via ipython
psutil==7.0.0
    # via ipykernel
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   fastapi
    #   google-genai
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via uv-example (./pyproject.toml)
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
    #   rich
pymongo==4.13.2
    # via uv-example (./pyproject.toml)
pymupdf==1.26.3
    # via uv-example (./pyproject.toml)
pypdfium2==4.30.1
    # via pdfplumber
pytest==8.4.1
    # via uv-example (./pyproject.toml)
python-dateutil==2.9.0.post0
    # via jupyter-client
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-magic==0.4.27
    # via uv-example (./pyproject.toml)
python-magic-bin==0.4.14
    # via uv-example (./pyproject.toml)
python-multipart==0.0.20
    # via fastapi
pywin32==310
    # via jupyter-core
pyyaml==6.0.2
    # via uvicorn
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
requests==2.32.4
    # via google-genai
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.8
    # via fastapi-cli
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
stack-data==0.6.3
    # via ipython
starlette==0.46.2
    # via
    #   uv-example (./pyproject.toml)
    #   fastapi
tenacity==8.5.0
    # via google-genai
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via openai
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typer==0.16.0
    # via fastapi-cli
typing-extensions==4.14.0
    # via
    #   anyio
    #   fastapi
    #   google-genai
    #   ipython
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via
    #   uv-example (./pyproject.toml)
    #   fastapi
    #   fastapi-cli
watchfiles==1.1.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
widgetsnbextension==4.0.14
    # via ipywidgets
